feat: gateway telegram extraction and event bridge supaplan_task:e42bf057-86c5-40a7-994d-711547e03e18
